requirements.txt
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.7
pathspec==0.9.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
tomli==1.0.4
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
